Split easy camera out of easy screenshot
